this took way to long to do...
sql script that creates the independent_films table. all sql databases use slightly different rules, and postgresql is no different so it took a while to get it to run smoothly. This data was gathered from what is available on the current website, and may need to be updated. Just needed something to work with while developing the gallery section of the films page